src/requirements.txt
Flask==3.1.1
PyJWT==2.10.1
python-dotenv==1.1.1

